+++ crates/tabby-common/src/registry.rs
@@ -74,6 +74,22 @@ impl ModelRegistry {
         models_dir().join(&self.name).join(name)
     }
 
+    pub fn migrate_model_path(&self, name: &str) -> Result<(), std::io::Error> {
+        let model_path = self.get_model_path(name);
+        let old_model_path = self
+            .get_model_dir(name)
+            .join(LEGACY_GGML_MODEL_RELATIVE_PATH);
+
+        if !model_path.exists() && old_model_path.exists() {
+            std::fs::rename(&old_model_path, &model_path)?;
+            #[cfg(target_family = "unix")]
+            std::os::unix::fs::symlink(&model_path, &old_model_path)?;
+            #[cfg(target_family = "windows")]
+            std::os::windows::fs::symlink_file(&model_path, &old_model_path)?;
+        }
+        Ok(())
+    }
+
     pub fn get_model_path(&self, name: &str) -> PathBuf {
         self.get_model_dir(name).join(GGML_MODEL_RELATIVE_PATH)
     }
@@ -104,4 +120,37 @@ pub fn parse_model_id(model_id: &str) -> (&str, &str) {
     }
 }
 
-pub static GGML_MODEL_RELATIVE_PATH: &str = "ggml/q8_0.v2.gguf";
+pub static LEGACY_GGML_MODEL_RELATIVE_PATH: &str = "ggml/q8_0.v2.gguf";
+pub static GGML_MODEL_RELATIVE_PATH: &str = "ggml/model.gguf";
+
+#[cfg(test)]
+mod tests {
+    use temp_testdir::TempDir;
+
+    use super::{ModelRegistry, *};
+    use crate::path::set_tabby_root;
+
+    #[tokio::test]
+    async fn test_model_migration() {
+        let root = TempDir::default();
+        set_tabby_root(root.to_path_buf());
+
+        let registry = ModelRegistry::new("TabbyML").await;
+        let dir = registry.get_model_dir("StarCoder-1B");
+
+        let old_model_path = dir.join(LEGACY_GGML_MODEL_RELATIVE_PATH);
+        tokio::fs::create_dir_all(old_model_path.parent().unwrap())
+            .await
+            .unwrap();
+        tokio::fs::OpenOptions::new()
+            .create(true)
+            .write(true)
+            .open(&old_model_path)
+            .await
+            .unwrap();
+
+        registry.migrate_model_path("StarCoder-1B").unwrap();
+        assert!(registry.get_model_path("StarCoder-1B").exists());
+        assert!(old_model_path.exists());
+    }
+}